refactor : 공통 위젯분리
